moved encoder wrapper class to util
