+++ src/main/java/frc/robot/util/Encoder.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package frc.robot.util;
 
 import com.revrobotics.RelativeEncoder;
